Move CAS1 Withdrawal Status Update Logic
This commit moves the logic to set the application status to withdrawn to the `Cas1ApplicationStatusService`, removing it from the `ApplicationListener`, which we want to remove.

The `ApplicationListener` is a remnant of when we used JPA listeners to update the status, which was hard to reason on, and was no longer supported in Spring Boot 3 / Hibernate 6 in the way we used them.