Revise application status when updated
Before this commit, once an unsubmitted application’s status was set to INAPPLICABLE it would never change to any other status, even if the unsubmitted application was updated to set inapplicable=false (e.g. the user pressed back and revised their answer).

This commit ensures the application status is recalculated whenever an unsubmitted application is updated, fixing the aforementioned bug.

This commit also removes the `ApplicationListener` as it has been replaced by the `Cas1ApplicationStatusService`.