Add `Cas1PlacementDates` container to PlacementApp
This commit introduces a `Cas1PlacementDates` class to contain placement dates for a Placement Application. Along with the function `PlacementApplicationEntity.placementDates()` this makes the nullable handling of expected arrival and duration cleaner, as either neither or both are provided.

Because this only returns dates for a submitted placement application, we’ve had to update various test data to make sure placement apps with dates are submitted, and any submitted placement apps have dates. This better reflects reality.