Fix placement application workload calculation
Before this change the calculation used for the following fields was incorrect wrt. placement applications:

* numTasksPending
* numTasksCompleted7Days
* numTasksCompleted30Days

The logic was using the placement applications’ ‘submitted_at’ date to determine if a placement application had been assessed. This is incorrect, and instead it should have used ‘decision_made_at’.

This commit fixes that logic and removes a redundant check on `decision` in the SQL as this is always set at the same time as `decision`