Add endpoint to get all premises capacities
This endpoint has been added as part of a spike to test performance when retrieving capacity for multiple premises. The endpoint will be further refined once user requirements are better understood.

Note that much like the existing capacity endpoint, the endpoint integration testing is limited as there is already detailed integration testing in the integration test for the `SpacePlanningService`