Use custom SQL to get OOSB for capacity
Before this commit we were retrieving `OutOfServiceBedEntity` when determining if there were out of service beds when calculating capacity.

Whilst this query was constrained to a given date range and premises, lazy loading was still required to find the latest out of service bed revision. This would make retreiving capacity for all premises very inefficient.

This commit changes the code to use custom SQL that only returns what is required, including information on the latest out of service bed revision.

The tests in `SpacePlanningServiceTest` have been improved to test various out of service bed scenarios to ensure the query is returning the expected values (e.g. not returning older out of service bed revision information)