Acquire pessimistic lock on CAS1 app update
Since enabled optimistic locking we’ve been seeing lock errors when updating applicaitons during the application drafting process. This commit acquires a pessimistic lock for this request. If the applicaiton is been submitted in the meantime (the reason we added the optimistic lock), this function will then error as it checks the application’s submission status

A more optimised solution would be to compare all fields in the update and if there are no changes return success (i.e. make the endpoint idempotent), because the lock is acquired