upstream /users/access returns requested CRNs
Before this commit the code assumed that calls to `/users/access` would not return an entry for a requested CRN if the CRN does not exist. This is not the case, and an entry for the CRN will always be returned.

This commit notes this on the client for this endpoint and updates the code to remove any code related to the CRN not existing