Remove bi-direction space booking transfer link
Despite being marked as lazy, the non-owning side of the transferredFrom/transferredTo `OneToOne` relationship was triggering eager loads. This is a common problem, dicussed here - https://thorben-janssen.com/hibernate-tip-lazy-loading-one-to-one/

The simplest solution to this problem was to make the relationship uni-directional, and use a query to find the ‘transferredFrom’ value in the few cases it is required.

This change was specifically made to optimise calculating capacity across multiple premises, which could involve 100s of space bookings.