Remove `SpaceBooking` planning model
This commit removes the `SpaceBooking` planning model and instead uses the `Cas1SpaceBookingEntity` returned from the repository. As part of this change, characteristics are no longer filtered on the Bed model to the 6 of interest for planning, which has no functional impact as the code consuming these characteristics in the `Cas1SpacePlanningService` already filters on the 6 characteristics of interest.