Manuall install Open Entity in View Interceptor
Open Entity (Session) In View (OSIV) is enabled by default in Spring, and whilst simplifying development wrt. lazy loading JPA relationships, it results in a database connection being held open from it’s first usage to completion of the API request. This can lead to exhausted connection pools under load, especially when calls are being made to upstream services in the same thread that holds the connection.

Because removing OSIV can cause subtle issues (especially are object equality), we can selectively disable it for certain paths that are causing particular issues. To support this we manually install the interceptor in the configuration class added by this commit.

This commit initially adds the interceptor to one endpoint as a proof of concept. We’ve confirmed locally that the interceptor is no longer being invoked for this specific endpoint