Don’t check user access if offender not LAO
Before this commit when calling `OffenderService.getPersonSummaryInfoResults` we would call the ‘user access’ endpoint to determine if the user could access the offender, even if we already knew there were no restrictions or exclusions on the offender for any users.

This commit changes that logic to only make upstream calls for CRNs where we know there is either a restriction or exclusion in place.