Update swagger groupings
This PR updates the swagger groupings, and adds missing files (which will be removed as part of https://dsdmoj.atlassian.net/browse/CAS-1845). This is to clean up the UI repo and remove non-cas2 models.

It also adds default documentation for the service name header (which was needed for a missing model)