Add validation to Nomis User
This adds a null check on the primary email address field from nomis user roles. This is a required field for sending emails, and raises errors in domain event processing, meaning messages go to the DLQ. This will prevent that from happening by ensuring the account always has an email address attached when the user logs in.