Fix LaoStrategy used in CAS1 Assessment Service
The `Cas1AssessmentService` is intended to replace the CAS1 elements of `AssessmentService`. When raising domain events, the `NeverRestricted` LAO strategy should be used, matching the behaviour in `AssessmentService`

This has highlighted some inefficient behaviour where we retrieve offender info twice - once with LAO Restrictions (to see if user has access) and then with no LAO restrictions (for the domain event).