Drop id prefixes from request for placement report
These prefixes were added to differentiate how a request for placement was persisted, but this differentiate is mostly academic to end users of the reports. Furthermore, we’re working on ‘fixing’ the request for placement model so the same table is used to represent all requests for placements. Also, the prefix `placement_request:` was misleading because the ID actually came from the `placement_applications_automatic` table

For this reason, this commit drops the id prefixes, providing just the UUID for each request for placement (which is derived from either `placement_applications.id` or `placement_applications_automatic.id`)