Add `placement_applications.automatic`
This will default to false, as all existing placement applications are not automatic.

This commit updates the queries used to return tasks to exclude any placement applications where automatic=true, because these will be created and approved automatically on application assessment.

Currently nothing is creating automatic placement applications, so the request for placement report does not need changing. This will follow in subsequent commits.