Rename placement_applications_automatic table
We now have the concept of an entry in `placement_applications` being automatic (via the ‘automatic’ boolean value)

To avoid confusion, this commit renames the `placement_applications_automatic` table to `placement_applications_placeholder`, as it really represents a limited set of information about a request for placement that will be ‘promoted’ to an entry in the `placement_applications` table once the corresponding application is approved. we need this placeholder to enable reporting on these values before the application has been approved.

In time, `placement_applications_placeholder` will be fully replaced by `placement_applications`.

Typically we’d avoid renaming tables ‘in place’ because it can lead to errors whilst rolling out the correspodning code change. As this table does not have any relationships in JPA and only used in two places in code, renaming in place is deemed acceptable.