Simplify PlacementRequirements creation
Before this commit the `createPlacementRequirements` function would return a `CasResult`, allowing for an error to be returned if the postcode isn’t recognised.

Given that this postcode is verified earlier in the process during creation of the original application, and cannot be modified after this, it’s fine to return an exception in this case (i.e. this should never happen). Futhermore, there’s nothing the user can do in response to this error because the postcode is locked in. They’d have to contact support anyway.

 This allows us to simplify the response type from `createPlacementRequirements`