Create ‘automatic’ placement app on assessment
This commit updates the assessment logic to automatically create `placement_applications` entry when an application with an implicit request for placement is approved.

At this point we also archive the corresponding `placement_applications_placeholder` entry, which was added to ensure we report on requests for placements the momnent the application is submitted. The report will now pick up on the new `placement_applications` entry so the placeholder is no longer required. The end goal is to complete remove `placement_applications_placeholder` and create an entry in `placemnent_applications` when the application is submitted.

We have checked that there is a corresponding entry in `placement_applications_placeholder` for every application pending approval using the following:

```
select count(*)
from approved_premises_applications apa
left outer join placement_applications_placeholder placeholder on placeholder.application_id = apa.id
inner join assessments on assessments.application_id = apa.id
where
apa.arrival_date is not null and
assessments.decision is null and
assessments.reallocated_at is null and
placeholder.id is null;
-- success --
count
-----
    0
```

A small update was required to the reports tests as ‘standard’ requests for placements are now represented by placement applications once the application has been assessed.