Remove seed jobs used for find and book rollout
This commit removes the 3 seed jobs used for the find and book rollout, and the cas1_delius_booking_import table, a temporary table used during rollout.

This table is safe to remove along with the code because it’s only used within a seed job that is no longer used.