Remove PlacementRequestEntity.reallocatedAt
This was used when placement requests were managed as tasks. This has not been the case for over a year, so this column can be removed.

Before making this change we deleted all existing realloacted placement requests, meaning all entries in prod have `reallocated_at = null`, making the field redundant.