Temporarily remove probation integration test user
The `AutomatedTestUser` is used by probation integration end to end tests. This is currently causing an issue as it’s using an unreognized probation region. To unblock local developement this commit removes the user from the seed file. Note that this user will still exist in dev/test environments and was added in this file for completeness.