Update assessment annotations
This moves the annotations into the actual controller and removes the previously committed auto-generated files.

This is still split between CAS's, so the annotations are just copied and pasted, and not refactored (to @GetMapping etc). This can be in a future commit when the classes are split into cas specific ones.