Add requested and authorised durations on RfPs
Request for placements currently capture a single `duration` value. For an ‘automatic’ request for placement it’s possible for the requested duration to differ from the authorised duration.

This commit introduces separate fields for these two durations. Currently they’re always populated from the existing ‘duration’ fields, which is required until we can backfill the ‘requestedDuration’ values which will follow in a subsequent commit