CAS-1795 Void Bedspace Cancellations V2
This PR creates the void bedspaces cancellations endpoint. It's based on the existing /lost-beds/cancellations.

As cancellations have been combined in to the void bedspace model, this has been updated from POST to PUT. The path has also been updated to include the bedspace ID, which will be updated to the existing endpoints in a future PR.