Remove duplicate request for placement domain evts
Now we’re creating a `placement_application` to represent the request for placement in the original application, we should ensure the `Request for Placement Created` domain event is linking to the placement applicaiton instead of the placement request i.e. in the domain event

```requestForPlacementId = placementApplication.id```

instead of

```requestForPlacementId = placementRequest.id```

This commit updates the aforementioned logic to use the existing `Cas1PlacementApplicationDomainEventService. placementApplicationSubmitted` function to create the domain event for initial requests, which links it to the placement application.

We then  completely remove the `Cas1PlacementRequestDomainEventService. placementRequestCreated` function, which was used to create a ‘Request for Placement Created’ domain event against the placement request for initial request for placements, but now would result in a duplicate domain event being raised (except the `requestForPlacementId` would be the `placementRequest.id`)

This commit also adds various notes where we can refactor the code to make it more cohesive and easier to read. I’ve not done that in this commit/PR as it’s already quite large.