Remove legacy test.
This is a legacy test from when the same endpoint was being used by both services.  It was testing that a CAS 1 user would be forbidden from deallocating a CAS 1 assessment, if CAS 3 X-Service-Name was sent in the request.  No longer applicable since the services have different allocate/deallocate end points now.