Remove use of deprecated func in PeopleController
This requires some updates to the integration tests because the new API interactions are slightly different.

The deprecated get offender endpoint would return a 404 if the offender isn’t found. Because we’re now using a bulk endpoint we need to configure it to not return anything when the CRN is requested (404 would indicate client misconfiguration)

Some endpoints previously used the ‘getOffender’ function to determine if a user can access the offender, which would retrieve the offender details and then make a second call to determine if the user could access them. The code now just makes that second call. That second call own’t tell the user if. the offender doesn’t exist, and will just return permission to access the offender. We previously had ‘offender not found’ integration tests for these requests which are no longer applicable. These have been replaced with ‘permission to access’ tests, which give better coverage.