Remove deprecated upstream calls in Cas1 Booking
This updates the `Cas1BookingDomainEventService` to stop using the deprecated function on `OffenderService` to retrieve the offenders NOMS number. In doing this we’ve simplified the code to remove duplication.

This update also removes some inconsistent behaviour around LAOs. Because we’re only retrieving the NOMS number to add to a domain event, LAO restrictions are ignored. This was already the case for most of the code in this class.

Some integration tests required updating to ensure the CRN on the space booking matched that on the applicaiton (as is always the case where an application is set). We also needed to use ‘actual’ offenders to ensure the new endpoint mocks were correctly setup