Retrieve keyworker details by staff code
Before this commit when retreiving keyworker details to persist on a booking we would use the staff code and premises code. This ensured the keyworker we were assigning belonged to the associated premises.

As we’re changing to allow any FUTURE_MANAGER to be assigned as a keyworker this restriction no longer makes sense, and we don’t want to be restricted by delius user to premises mappings.

This commit updates the logic to retrieve keyworker information using staff code only.