Support setting keywork using user id
This commit allows a keyworker to be assigned using staffCode or a CAS1 user ID.

Using the staff code is the current behaviour and should continue to be used in production. If a user can’t be found for a staff code a warning will be logged. This is acceptable because we’re currently still using the staff code in production when retreiving keyworker information.

Using the CAS1 User ID is new behaviour that has been added to support development of new assign keyworker functionality in environments outside of production.

Eventually all keyworkers will be assigned using the CAS1 User ID, at which point support for setting key workers using the staff ID can be removed, simplifying this logic.