Only allow future managers in cas1 staff list
The endpoint `/cas1/premises/{premisesId}/staff` is used to retrieve a list of staff members from delius that can be used in CAS1 as key workers.

As part of our move to use the user’s table to manage the list of potential keyworkers, we want to move away from using this endpoint.

In the meantime, Once we’ve backfilled the new `cas1_space_bookings.key_worker_user_id` for all records in production we want to ensure that only staff members with a correspoding entry in the `users` table with the role `CAS1_FUTURE_MANAGER` can be assigned as key workers.

This commit updates `/cas1/premises/{premisesId}/staff` to ensure this is the case (and will only be merged/deployed once the user id backfill is complete, and we can ensure all keyworkers have the `CAS1_FUTURE_MANAGER` role)