Tidy up use of UserQualification
This commit moves the logic used to determine what qualifications are required to assess an application from the `ApplicationEntity` class into a CAS1 specific `Cas1AssessmentService`. This then allows us to remove the abstract ‘getRequiredQualifications’ function on `ApplicationEntity`.

This highlights that there is cas1 specific task reallocation code in `AssessmentService` that should really be moved to `Cas1AssessmentService`

As part of this change documentation has been added onto each qualifcation