Refactor Cas3 V2 void bedspaces
This is a further refactor the v2 void bedspaces. The 'new' and 'update' models were similar, with the exception of bedspaceId. Since this is now included as part of the url, it's not required in the models, so they have been combined into one. This also meant the integration tests could be simplified.

It also creates the cancellations model, as the only field used from the Cas3VoidBedsapce model was cancellation notes.

Also renaming and standarising integration test names.