formatting
Remove @Schema annotations.

This PR is the first to start removing the @Schema annotations that are leftover from the autogenerated files. This first PR just removed the annotations that do not contain any other information. The rest will be in another PR, with any relevant comments updated to use the @Param annotations.

This will have the benefit of removing 1000's of line of "This annotation is currently applied to the value parameter only, but in the future it will also be applied to field." comments in the logs